Scrap the Rust code
I decided that I will switch for Rust for this project to C++. I chose
to do this because a lot of the compiler 'quirks' are just plain
annoying. For example, a macro must end with !. Why? The compiler can
tell you about this error, so why doesn't it just ignore it? What's the
difference between a function and macro anyways? A lot of this could be
answered by simply having more knowledge and experience, but I find it
annoying. I wonder how those gits in servo/servo do it.